fix(python3): various fixes for Python 3
This is a variety of different fixes required for Python. Among them are
replacing most references to `unicode` with `str` and `long` to `int`. This
also corrects some of the exception handling to make the it Python 3 compatible
and fixes the date parser for XMP data.

+++ src/pyexiv2/exif.py
@@ -312,7 +312,7 @@ class ExifTag(ListenerInterface):
 
         elif self.type in ('Long', 'SLong'):
             try:
-                return long(value)
+                return int(value)
             except ValueError:
                 raise ExifValueError(value, self.type)
